+++ test/Fixture/Definitions/DoesNotImplementDefinition/RepositoryDefinition.php
@@ -21,6 +21,6 @@ final class RepositoryDefinition
 {
     public function accept(FixtureFactory $fixtureFactory, Generator $faker): void
     {
-        $fixtureFactory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
+        $fixtureFactory->define(Fixture\FixtureFactory\Entity\Repository::class);
     }
 }