+++ test/Fixture/Definitions/ImplementsDefinition/RepositoryDefinition.php
@@ -22,6 +22,6 @@ final class RepositoryDefinition implements Definition
 {
     public function accept(FixtureFactory $fixtureFactory, Generator $faker): void
     {
-        $fixtureFactory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
+        $fixtureFactory->define(Fixture\FixtureFactory\Entity\Repository::class);
     }
 }