+++ test/Fixture/Definitions/ImplementsDefinitionButHasPrivateConstructor/RepositoryDefinition.php
@@ -26,6 +26,6 @@ final class RepositoryDefinition implements Definition
 
     public function accept(FixtureFactory $fixtureFactory, Generator $faker): void
     {
-        $fixtureFactory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
+        $fixtureFactory->define(Fixture\FixtureFactory\Entity\Repository::class);
     }
 }